Changes applied after pull request
