+++ FizzBuzz/FizzBuzz/Program.cs
@@ -5,66 +5,27 @@ using System.Linq;
 
 namespace FizzBuzz
 {
-    public class Flags
-    {
-        public bool three = false;
-        public bool five = false;
-        public bool seven = false;
-        public bool eleven = false;
-        public bool thirteen = false;
-        public bool seventeen = false;
-    }
-
     public class Program
     {
-        private Flags setFlags(string[] args)
-        {
-            Flags flags = new Flags();
-            for (int i = 0; i < args.Length; i++)
-            {
-                if (args[i] == "3")
-                    flags.three = true;
-                else if (args[i] == "7")
-                    flags.seven = true;
-                else if (args[i] == "5")
-                    flags.five = true;
-                else if (args[i] == "11")
-                    flags.eleven = true;
-                else if (args[i] == "13")
-                    flags.thirteen = true;
-                else if (args[i] == "17")
-                    flags.seventeen = true;
-            }
-            return flags;
-        }
-
-        private string getIterations()
-        {
-            Console.WriteLine("Write maximum number : ");
-            string max;
-            max = Console.ReadLine();
-            return max;
-        }
-
-        public string FizzBuzzLogic(int i, Flags flags)
+        public string FizzBuzzLogic(int i, Rule Rule)
         {
             string message = "";
-            if (i % 3 == 0 && flags.three)
+            if (i % 3 == 0 && Rule.three)
             {
-                if (i % 5 == 0 && flags.five)
+                if (i % 5 == 0 && Rule.five)
                     message = "FizzBuzz";
                 else
                     message = "Fizz";
             }
-            else if (i % 5 == 0 && flags.five)
+            else if (i % 5 == 0 && Rule.five)
                 message = "Buzz";
-            if (i % 7 == 0 && flags.seven)
+            if (i % 7 == 0 && Rule.seven)
                 if (String.IsNullOrEmpty(message))
                     message = "Bang";
                 else message += "Bang";
-            if (i % 11 == 0 && flags.eleven)
+            if (i % 11 == 0 && Rule.eleven)
                 message = "Bong";
-            if (i % 13 == 0 && flags.thirteen)
+            if (i % 13 == 0 && Rule.thirteen)
                 if (String.IsNullOrEmpty(message))
                     message = "Fezz";
                 else if (!message.Contains('B'))
@@ -76,7 +37,7 @@ namespace FizzBuzz
                     string second_message = message.Substring(index, message.Length - first_message.Length);
                     message = first_message + "Fezz" + second_message;
                 }
-            if (i % 17 == 0 && !String.IsNullOrEmpty(message) && flags.seventeen)
+            if (i % 17 == 0 && !String.IsNullOrEmpty(message) && Rule.seventeen)
             {
                 string reversed_message = "";
                 int pos = message.Length;
@@ -97,19 +58,19 @@ namespace FizzBuzz
         static void Main(string[] args)
         {
             var program = new Program();
-            Flags flags = program.setFlags(args);
+            Rule Rule = program.setRule(args);
             string max = program.getIterations();
 
             for (int i = 1; i <= Int64.Parse(max); i++)
             {
-                string message = program.FizzBuzzLogic(i, flags);
+                string message = program.FizzBuzzLogic(i, Rule);
 
                 Console.WriteLine(message);
             }
 
             Console.WriteLine("Using IEnumerable ... ");
 
-            var fizzBuzzer = new FizzBuzz(max, flags, program);
+            var fizzBuzzer = new FizzBuzz(max, Rule, program);
 
             foreach (var value in fizzBuzzer)
             {
@@ -117,33 +78,6 @@ namespace FizzBuzz
             }
 
             Console.WriteLine("One-Liner ... ");
-
-            //Fizz - > 3
-            //Buzz - > 5
-            //Bang - > 7
-
-            //rewritten code more suitable for inline ifs : 
-
-            //if (i % 3 == 0)
-            //{
-            //    if (i % 5 == 0 && i % 7 == 0)
-            //        message = "FizzBuzzBang";
-            //    else  if (i % 5 == 0)
-            //        message = "FizzBuzz";
-            //    else if (i % 7 == 0)
-            //          message = "FizzBang"
-            //    else
-            //        message = "Fizz";
-            //}
-            //else if (i % 5 == 0)
-            //  { if (i % 7 == 0)
-            //          message = "BuzzBang";
-            //    else
-            //          message = "Buzz"
-            //  }
-            //else if (i % 7 == 0)
-            //    message = "Bang";
-            //else message = i.ToString();
 
             IEnumerable fizzbuzzED = Enumerable.Range(1, Int32.Parse(max)).Select(x => x % 3 == 0 ? (x % 5 == 0 && x % 7 == 0 ? "FizzBuzzBang" : (x % 5 == 0 ? "FizzBuzz" : (x % 7 == 0 ? "FizzBang" : "Fizz"))) : (x % 5 == 0 ? (x % 7 == 0 ? "BuzzBang" : "Buzz") : (x % 7 == 0 ? "Bang" : x.ToString())));
 
@@ -153,74 +87,44 @@ namespace FizzBuzz
             }
 
         }
-    }
-
-    public class FizzBuzz : IEnumerable
-    {
-        private string[] fizzbuzzED;
-
-        public FizzBuzz(string max, Flags flags, Program program)
+        private Rule setRule(string[] args)
         {
-            fizzbuzzED = new string[Int64.Parse(max)];
-
-            for (int i = 1; i <= Int64.Parse(max); i++)
+            Rule Rule = new Rule();
+            for (int i = 0; i < args.Length; i++)
             {
-                string message = program.FizzBuzzLogic(i, flags);
-
-                fizzbuzzED[i - 1] = message;
-            }
-        }
-
-        public IEnumerator GetEnumerator()
-        {
-            return new FizzBuzzEnumerator(fizzbuzzED);
-        }
-    }
-
-    public class FizzBuzzEnumerator : IEnumerator
-    {
-        public string[] fizzbuzzED;
-
-        int position = -1;
-
-        public FizzBuzzEnumerator(string[] list)
-        {
-            fizzbuzzED = list;
-        }
-
-        public bool MoveNext()
-        {
-            position++;
-            return (position < fizzbuzzED.Length);
-        }
-
-        public void Reset()
-        {
-            position = -1;
-        }
-
-        object IEnumerator.Current
-        {
-            get
-            {
-                return Current;
-            }
-        }
-
-        public string Current
-        {
-            get
-            {
-                try
+                if (args[i] == "3")
                 {
-                    return fizzbuzzED[position];
+                    Rule.three = true;
                 }
-                catch (IndexOutOfRangeException)
+                else if (args[i] == "7")
                 {
-                    throw new InvalidOperationException();
+                    Rule.seven = true;
+                }
+                else if (args[i] == "5")
+                {
+                    Rule.five = true;
+                }
+                else if (args[i] == "11")
+                {
+                    Rule.eleven = true;
+                }
+                else if (args[i] == "13")
+                {
+                    Rule.thirteen = true;
+                }
+                else if (args[i] == "17")
+                {
+                    Rule.seventeen = true;
                 }
             }
+
+            return Rule;
+        }
+
+        private string getIterations()
+        {
+            Console.WriteLine("Write maximum number : ");
+            return Console.ReadLine();
         }
     }
-
 }
