+++ FizzBuzz/FizzBuzz/Rule.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace FizzBuzz
+{
+    public class Rule
+    {
+        public bool three = false;
+        public bool five = false;
+        public bool seven = false;
+        public bool eleven = false;
+        public bool thirteen = false;
+        public bool seventeen = false;
+    }
+}
